zaimplementowane metody GetMinimalCost, GetCosts z interfejsu

+++ branches/Pszczoly/Model.Bee/Bee.cs
@@ -9,10 +9,10 @@ namespace Model.Bee
     {
         public int status; // 0 = inactive, 1 = active, 2 = scout
         public int[] memoryMatrix; // problem-specific. a path of cities.
-        public double measureOfQuality; // smaller values are better. total distance of path.
+        public int measureOfQuality; // smaller values are better. total distance of path.
         public int numberOfVisits;
 
-        public Bee(int status, int[] memoryMatrix, double measureOfQuality, int numberOfVisits)
+        public Bee(int status, int[] memoryMatrix, int measureOfQuality, int numberOfVisits)
         {
             this.status = status;
             this.memoryMatrix = new int[memoryMatrix.Length];
